enemy, armor, health, equipment improvements

+++ Attack enemys 2D rpg style/Assets/Scripts/Player/CoinCounter.cs
@@ -19,7 +19,7 @@ public class CoinCounter : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        coinAmount = PlayerScr.Gold;
+        coinAmount = (int)PlayerScr.Gold;
         text.text = coinAmount.ToString();
     }
 }
